requirements.txt
nose>=1.2.1
ddt>=0.8.0
jsonpickle==0.4.0
requests==2.0.1
wsgiref==0.1.2
reconfigure==0.1.32
regd==1.3.1b
funcy>=0.5.6
selenium==2.36.0
furl==0.3.6
redis
funky
thrift
elasticsearch
gitpython==1.0.0
clipboard
bernhard
Appium-Python-Client

# pyautoit - only for Windows

# Windows: There may be problems that need binary installation (use wheel and binary packages)
#pika==0.9.13
#blist
#paramiko==1.7.4
#pycrypto==2.6
#psycopg2==2.5.4
#MySQL-python==1.2.4

# deprecate
# cassandra-driver